server/requirements.txt
flask>=3.1.0
flask-cors>=5.0.1
flask-sqlalchemy>=3.1.1
psycopg2-binary>=2.9.10
python-dotenv>=1.0.1
gunicorn>=21.2.0
pytest>=7.4.0
pytest-flask>=1.2.0

